working frontend and backend on course methods

+++ backend/src/courses/courses.service.ts
@@ -11,7 +11,10 @@ export class CoursesService {
   ) {}
 
   findAll(): Promise<Course[]> {
-    return this.coursesRepository.find();
+    return this.coursesRepository
+      .createQueryBuilder('course')
+      .orderBy('course.position', 'ASC')
+      .getMany();
   }
 
   findAllValid(): Promise<Course[]> {
@@ -26,11 +29,46 @@ export class CoursesService {
     return this.coursesRepository.save(course);
   }
 
-  async update(id: number, course: Course): Promise<void> {
+  async update(id: number, course: Course): Promise<Course | null> {
     await this.coursesRepository.update(id, course);
+    await this.updateAllValidCoursesPositions(course);
+
+    return this.coursesRepository.findOne({ where: { id } });
   }
 
-  async remove(id: number): Promise<void> {
-    await this.coursesRepository.delete(id);
+  async remove(id: number): Promise<boolean> {
+    const course = await this.coursesRepository.findOne({ where: { id } });
+    if (course) {
+      const deleted = await this.coursesRepository.delete(id);
+      if (course.valid) {
+        await this.updateAllValidCoursesPositions(course);
+      }
+      return !!deleted.affected;
+    }
+    return false;
+  }
+
+  // This method updates the positions of all valid courses
+  async updateAllValidCoursesPositions(updatedCourse: Course): Promise<void> {
+    const validCourses = await this.findAllValid();
+    const updatedCourseIndex = validCourses.findIndex(
+      (course) => course.id === updatedCourse.id,
+    );
+
+    if (updatedCourseIndex !== -1) {
+      validCourses.splice(updatedCourseIndex, 1);
+    }
+
+    if (updatedCourse.position === null) {
+      validCourses.push(updatedCourse); // Add to the end if position is null
+    } else {
+      validCourses.splice(updatedCourse.position - 1, 0, updatedCourse);
+    }
+
+    for (let i = 0; i < validCourses.length; i++) {
+      validCourses[i].position = i + 1;
+    }
+
+    await this.coursesRepository.save(validCourses);
   }
 }
